Adapt changes to new ParallelRunnable trait and cleanup
